Works with XEM6310, somewhat...
In the previous commit, I used the SDRAMFifo.v from Intan's Rhythm
Interface directly with minimal modifications. I took the component
modules from the MIG/ folder of Opal Kelly's RAMTester sample for
XEM6310, changed the wire/reg widths to fit SDRAMFifo and called it a
day.

However, this is not good because the FrontPanel communication for USB2
and USB3 is different. Specifically this means two things:

1) Pipe data is transfered through USB3 in 8 bits, but transferred
to/from FPGA in 32-bits.

  Pipe data is transferred through USB2 in 8 bits, but transferred
to/from FPGA in 16-bits.

  This means the okPipeOut fifo needs to change so that its interface
with okHost is 32-bits instead of 16-bits.

  Along this vein, I modified main.v, SDRAMFifo.v, ddr2_state_machine.v,
the generated fifos, and the associated memory controllers to better match that used in
XEM6310's RAMTester sample code.

  SDRAMFifo.v referenced ramtest.v
  ddr2_state_machine.v referenced ddr2_test.v

  The fifos changes from:
        - okPipeIn: fifo_w16_2048_r64_512 -> fifo_w16_2048_r32_1024
        - okPipeOut: fifo_w64_512_r16_2048 -> fifo_w32_1024_r32_1024

2) Because of the fifos aren't symmetric, have to change the 16-bits
word order when inputting to okPipeOut.

These files now bitgen successfully and can be uploaded onto the FPGA.
Running the IntanEvalBoard with this bitfile still gets the Incorrect
Header problem.

But this is only on some (most of) the readDataBlock() calls. And using
gdb it is clear the byte order output from the FPGA is correct, i.e. the
magic number is correct and the timestamp is present and in the correct
order.

TODO: Investigate whether this problem can be fixed by block throttle
output.