Tried using BTpipeout, does not solve the problem.
Increased the input/ouput fifos for the SDRAM, also did not solve the
problem.

However, there is progress. Seems like the magic number/frame-data
corruption only happens when I read in blocks more than 4576 bytes
at a time... need more tests to confirm this, but I wonder if this is
due to opal kelly's USB3 implementation. Their RAMTester reads 512 bytes
at a time, so this is definitely possible.

Need to think about this! Because open-ephys apparently is not doing
anything about this...so that's strange. Investigate!