It works!!!
The frequent glitch (of consistently 4 bytes) is a result of using "<="
rather than "<" in the comparison of ob_count to (FIFO_SIZE-1-BURST_LEN)
in ddr2_state_machine's s_idle. This may cause overflow of the output
FIFO by 1 32-bit words.

TODO: With this removed, do combinations of BURST_LEN and BLOCK_SIZE
still cause failures in piperead ever?